Split MemoryBlock into an abstract class and two inheritors
