+++ BizHawk.Common/BizInvoke/CallingConventionAdapter.cs
@@ -140,7 +140,9 @@ namespace BizHawk.Common.BizInvoke
 			public SysVHostMsGuest()
 			{
 				int size = 4 * 1024 * 1024;
-				_memory = new MemoryBlock((ulong)size);
+				_memory = PlatformLinkedLibSingleton.RunningOnUnix
+					? (MemoryBlock) new MemoryBlockUnix((ulong) size)
+					: (MemoryBlock) new MemoryBlockWin32((ulong) size);
 				_memory.Activate();
 				_refs = new WeakReference[size / BlockSize];
 			}
