+++ BizHawk.Emulation.Cores/Waterbox/MapHeap.cs
@@ -1,4 +1,5 @@
-﻿using BizHawk.Emulation.Common;
+﻿using BizHawk.Common;
+using BizHawk.Emulation.Common;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -50,7 +51,9 @@ namespace BizHawk.Emulation.Cores.Waterbox
 		public MapHeap(ulong start, ulong size, string name)
 		{
 			size = WaterboxUtils.AlignUp(size);
-			Memory = new MemoryBlock(start, size);
+			Memory = PlatformLinkedLibSingleton.RunningOnUnix
+				? (MemoryBlock) new MemoryBlockUnix(start, size)
+				: (MemoryBlock) new MemoryBlockWin32(start, size);
 			Name = name;
 			_pagesAsBytes = new byte[size >> WaterboxUtils.PageShift];
 			_pages = (MemoryBlock.Protection[])(object)_pagesAsBytes;
