+++ BizHawk.Emulation.Cores/Waterbox/PeWrapper.cs
@@ -172,7 +172,9 @@ namespace BizHawk.Emulation.Cores.Waterbox
 			// OK, NOW MOUNT
 
 			LoadOffset = (long)Start - (long)_pe.ImageNtHeaders.OptionalHeader.ImageBase;
-			Memory = new MemoryBlock(Start, Size);
+			Memory = PlatformLinkedLibSingleton.RunningOnUnix
+				? (MemoryBlock) new MemoryBlockUnix(Start, Size)
+				: (MemoryBlock) new MemoryBlockWin32(Start, Size);
 			Memory.Activate();
 			Memory.Protect(Start, Size, MemoryBlock.Protection.RW);
